refactor: restructure backend into feature-based modules (stigmata/weapons/astralops), flatten weapon schema, add OpenAPI response validation and error messages, extract checkAuth utility
