+++ src/backend/db/schema.ts
@@ -63,17 +63,10 @@ export type SelectStigmata = typeof stigmata.$inferSelect & {
 export const weapon = sqliteTable("weapon", {
   id: integer("id").primaryKey({ autoIncrement: true }),
   name: text("name").notNull(),
-  atk: integer("atk"),
-  crt: integer("crt"),
-})
-
-export const weaponImages = sqliteTable("weapon_images", {
-  id: integer("id").primaryKey({ autoIncrement: true }),
-  weaponId: integer("weapon_id")
-    .notNull()
-    .references(() => weapon.id),
-  baseUrl: text("base_url"),
-  maxUrl: text("max_url"),
+  atk: integer("atk").notNull(),
+  crt: integer("crt").notNull(),
+  baseUrl: text("base_url").notNull(),
+  maxUrl: text("max_url").notNull(),
 })
 
 export const weaponSkills = sqliteTable("weapon_skills", {
@@ -87,7 +80,6 @@ export const weaponSkills = sqliteTable("weapon_skills", {
 
 export type InsertWeapon = typeof weapon.$inferInsert
 export type SelectWeapon = typeof weapon.$inferSelect & {
-  images: (typeof weaponImages.$inferSelect)[]
   skills: (typeof weaponSkills.$inferSelect)[]
 }
