+++ src/backend/modules/index.ts
@@ -1,3 +1,3 @@
 export { stigmataRoutes, getStigmata } from "./stigmata"
 export { weaponRoutes, getWeapon } from "./weapons"
-export { astralOpRoutes, getAstralOp } from "./astralops"
+export { astralOpRoutes, getAstralOp } from "./astralOps"
