--- src/examples/post_stigmata.json
@@ -1,103 +1,0 @@
-{
-  "Single Stigmata Post": {
-    "name": "Test Stigma",
-    "positions": [
-      {
-        "position": "B",
-        "name": "Test Stigma (B)",
-        "skillName": "Test Skill",
-        "skillDescription": "Test Skill Description",
-        "stats": {
-          "hp": 0,
-          "atk": 0,
-          "def": 0,
-          "crt": 0,
-          "sp": 0
-        }
-      }
-    ],
-    "images": [
-      {
-        "position": "B",
-        "iconUrl": "public/img/stigmata/MEI_B_icon.png",
-        "bigUrl": "public/img/stigmata/MEI_B_big.png"
-      }
-    ],
-    "setEffects": {
-      "setName": "Test Set Effect",
-      "twoPieceName": "Two Test",
-      "twoPieceEffect": "Two Test Effect",
-      "threePieceName": "Three Test",
-      "threePieceEffect": "Three Test Effect"
-    }
-  },
-
-  "All Positions Post": {
-    "name": "Test Stigmata",
-    "positions": [
-      {
-        "position": "T",
-        "name": "Test Stigmata (T)",
-        "skillName": "Test Skill (T)",
-        "skillDescription": "Test Skill Description (T)",
-        "stats": {
-          "hp": 0,
-          "atk": 0,
-          "def": 0,
-          "crt": 0,
-          "sp": 0
-        }
-      },
-      {
-        "position": "M",
-        "name": "Test Stigmata (M)",
-        "skillName": "Test Skill (M)",
-        "skillDescription": "Test Skill Description (M)",
-        "stats": {
-          "hp": 0,
-          "atk": 0,
-          "def": 0,
-          "crt": 0,
-          "sp": 0
-        }
-      },
-      {
-        "position": "B",
-        "name": "Test Stigmata (B)",
-        "skillName": "Test Skill (B)",
-        "skillDescription": "Test Skill Description (B)",
-        "stats": {
-          "hp": 0,
-          "atk": 0,
-          "def": 0,
-          "crt": 0,
-          "sp": 0
-        }
-      }
-    ],
-    "images": [
-      {
-        "position": "T",
-        "iconUrl": "public/img/stigmata/MEI_T_icon.png",
-        "bigUrl": "public/img/stigmata/MEI_T_big.png"
-      },
-      {
-        "position": "M",
-        "iconUrl": "public/img/stigmata/MEI_M_icon.png",
-        "bigUrl": "public/img/stigmata/MEI_M_big.png"
-      },
-      {
-        "position": "B",
-        "iconUrl": "public/img/stigmata/MEI_B_icon.png",
-        "bigUrl": "public/img/stigmata/MEI_B_big.png"
-      }
-    ],
-    "setEffects": {
-      "setName": "Test Set Effect",
-      "twoPieceName": "Two Test",
-      "twoPieceEffect": "Two Test Effect",
-      "threePieceName": "Three Test",
-      "threePieceEffect": "Three Test Effect"
-    }
-  }
-}
